Finish 5_moving_average.c
In this assignment, I wrote two versions of moving average one
in O(NM) and the other is in O(N). I use result of O(NM) to test
the result of O(N) because I think the method to write O(NM) is
easier and unlikely to have bug in there.